Refactor age to reuse pages and questions
