--- app/models/form/sales/pages/person4_age.rb
@@ -1,19 +1,0 @@
-class Form::Sales::Pages::Person4Age < ::Form::Page
-  def initialize(id, hsh, subsection)
-    super
-    @id = "person_4_age"
-    @header = ""
-    @description = ""
-    @subsection = subsection
-    @depends_on = [
-      { "details_known_4" => 1 },
-    ]
-  end
-
-  def questions
-    @questions ||= [
-      Form::Sales::Questions::Person4AgeKnown.new(nil, nil, self),
-      Form::Sales::Questions::Person4Age.new(nil, nil, self),
-    ]
-  end
-end
