+++ spec/models/form/sales/subsections/household_characteristics_spec.rb
@@ -36,20 +36,24 @@ RSpec.describe Form::Sales::Subsections::HouseholdCharacteristics, type: :model 
         number_of_others_in_property
         person_1_known
         person_1_age
+        person_1_age_joint_purchase
         person_1_gender_identity
         person_1_gender_identity_joint_purchase
         person_1_working_situation
         person_1_working_situation_joint_purchase
         person_2_known
         person_2_age
+        person_2_age_joint_purchase
         person_2_working_situation
         person_2_working_situation_joint_purchase
         person_3_known
         person_3_age
+        person_3_age_joint_purchase
         person_3_working_situation
         person_3_working_situation_joint_purchase
         person_4_known
         person_4_age
+        person_4_age_joint_purchase
         person_4_working_situation
         person_4_working_situation_joint_purchase
       ],
